Fixes issue #59
This turned out to be fairly simple on the whole. The basic fix was to add some better processing in XSLT. I do not have a working test environment locally, so I will be double checking this on the dev machine.

Changes in DocketXML also fix an issue where a mutable variable is provided as a function arguemtnt. I did not analyze what real-world problems might have been caused by this, but in my experience this is a nasty mistake to make because it causes the value from the previous usage of a function to linger. The common way this happens is with a statement like:

 def my_function(var={}):

 Because var is a dict, it is mutable and will cause trouble. The better way to do this is:

 def my_function(var=None):
     if var is None:
         var = {}

 And so I've converted this where necessary.

 Also, lots of updates to the XSL formatting.